refactor: improve error handling (#25)

+++ internal/ui/render_helpers.go
@@ -11,16 +11,16 @@ func (issue *Issue) setDesc() {
 	issueType := getIssueTypeStyle(issue.issueType).Render(issue.issueType)
 
 	if issue.assignee != "" {
-		assignee = assigneeStyle(issue.assignee).Render(RightPadTrim(issue.assignee, int(listWidth/4)))
+		assignee = assigneeStyle(issue.assignee).Render(RightPadTrim(issue.assignee, listWidth/4))
 	} else {
-		assignee = assigneeStyle(issue.assignee).Render(RightPadTrim("", int(listWidth/4)))
+		assignee = assigneeStyle(issue.assignee).Render(RightPadTrim("", listWidth/4))
 	}
 
-	status = issueStatusStyle.Render(RightPadTrim(issue.status, int(listWidth/4)))
+	status = issueStatusStyle.Render(RightPadTrim(issue.status, listWidth/4))
 
 	if issue.aggSecondsSpent > 0 {
 		totalSecsSpent = aggTimeSpentStyle.Render(humanizeDuration(issue.aggSecondsSpent))
 	}
 
-	issue.desc = fmt.Sprintf("%s%s%s%s%s", RightPadTrim(issue.issueKey, int(listWidth/4)), status, assignee, issueType, totalSecsSpent)
+	issue.desc = fmt.Sprintf("%s%s%s%s%s", RightPadTrim(issue.issueKey, listWidth/4), status, assignee, issueType, totalSecsSpent)
 }
